+++ packages/alfa-css/test/value/calculation.spec.ts
@@ -1,12 +1,10 @@
 import { test } from "@siteimprove/alfa-test";
 
 import { Lexer } from "../../src/syntax/lexer";
-import { Calculation } from "../../src/value/calculation";
+import { Math } from "../../src/value/math-expression";
 
 function parse(input: string) {
-  return Calculation.parse(Lexer.lex(input)).map(
-    ([, calculation]) => calculation
-  );
+  return Math.parse(Lexer.lex(input)).map(([, calculation]) => calculation);
 }
 
 test(".parse() parses an addition expression of numbers", (t) => {
@@ -15,13 +13,10 @@ test(".parse() parses an addition expression of numbers", (t) => {
   t(calculation.isNumber());
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "number",
-        value: 3,
-      },
+      value: { type: "number", value: 3 },
     },
   });
 });
@@ -32,13 +27,10 @@ test(".parse() parses an addition expression of percentages", (t) => {
   t(calculation.isPercentage());
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "percentage",
-        value: 0.03,
-      },
+      value: { type: "percentage", value: 0.03 },
     },
   });
 });
@@ -49,14 +41,10 @@ test(".parse() parses an addition expression of absolute lengths", (t) => {
   t(calculation.isDimension("length"));
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "length",
-        value: 3,
-        unit: "px",
-      },
+      value: { type: "length", value: 3, unit: "px" },
     },
   });
 });
@@ -67,14 +55,10 @@ test(".parse() parses an addition expression of relative lengths", (t) => {
   t(calculation.isDimension("length"));
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "length",
-        value: 3,
-        unit: "em",
-      },
+      value: { type: "length", value: 3, unit: "em" },
     },
   });
 });
@@ -85,25 +69,22 @@ test(".parse() parses an addition expression of mixed lengths", (t) => {
   t(calculation.isDimension("length"));
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
-      type: "sum",
-      operands: [
+      type: "calculation",
+      arguments: [
         {
-          type: "value",
-          value: {
-            type: "length",
-            value: 1,
-            unit: "px",
-          },
-        },
-        {
-          type: "value",
-          value: {
-            type: "length",
-            value: 2,
-            unit: "em",
-          },
+          type: "sum",
+          operands: [
+            {
+              type: "value",
+              value: { type: "length", value: 1, unit: "px" },
+            },
+            {
+              type: "value",
+              value: { type: "length", value: 2, unit: "em" },
+            },
+          ],
         },
       ],
     },
@@ -116,24 +97,22 @@ test(".parse() parses an addition expression of a length and a percentage", (t) 
   t(calculation.isDimensionPercentage("length"));
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
-      type: "sum",
-      operands: [
+      type: "calculation",
+      arguments: [
         {
-          type: "value",
-          value: {
-            type: "length",
-            value: 1,
-            unit: "px",
-          },
-        },
-        {
-          type: "value",
-          value: {
-            type: "percentage",
-            value: 0.02,
-          },
+          type: "sum",
+          operands: [
+            {
+              type: "value",
+              value: { type: "length", value: 1, unit: "px" },
+            },
+            {
+              type: "value",
+              value: { type: "percentage", value: 0.02 },
+            },
+          ],
         },
       ],
     },
@@ -146,13 +125,10 @@ test(".parse() parses a multiplication expression of numbers", (t) => {
   t(calculation.isNumber());
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "number",
-        value: 6,
-      },
+      value: { type: "number", value: 6 },
     },
   });
 });
@@ -163,13 +139,10 @@ test(".parse() parses a multiplication expression of a number and a percentage",
   t(calculation.isPercentage());
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "percentage",
-        value: 0.06,
-      },
+      value: { type: "percentage", value: 0.06 },
     },
   });
 });
@@ -180,71 +153,52 @@ test(".parse() parses a multiplication expression of a number and a length", (t)
   t(calculation.isDimension("length"));
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "length",
-        value: 6,
-        unit: "px",
-      },
+      value: { type: "length", value: 6, unit: "px" },
     },
   });
 });
 
 test(".parse() gives higher precedence to * and / than + and -", (t) => {
   t.deepEqual(parse("calc(2 * 3 + 4)").get().toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "number",
-        value: 10,
-      },
+      value: { type: "number", value: 10 },
     },
   });
 
   t.deepEqual(parse("calc(2 + 3 * 4)").get().toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "number",
-        value: 14,
-      },
+      value: { type: "number", value: 14 },
     },
   });
 
   t.deepEqual(parse("calc(2 * 3 - 4)").get().toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "number",
-        value: 2,
-      },
+      value: { type: "number", value: 2 },
     },
   });
 
   t.deepEqual(parse("calc(3 / 2 + 4)").get().toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "number",
-        value: 5.5,
-      },
+      value: { type: "number", value: 5.5 },
     },
   });
 
   t.deepEqual(parse("calc(3 / 2 - 4)").get().toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "number",
-        value: -2.5,
-      },
+      value: { type: "number", value: -2.5 },
     },
   });
 });
@@ -255,13 +209,86 @@ test(".parse() parses a nested calc() function", (t) => {
   t(calculation.isNumber());
 
   t.deepEqual(calculation.toJSON(), {
-    type: "calculation",
+    type: "math expression",
     expression: {
       type: "value",
-      value: {
-        type: "number",
-        value: 6,
-      },
+      value: { type: "number", value: 6 },
+    },
+  });
+});
+
+test(".parse() parses longer operations chain", (t) => {
+  const calculation1 = parse("calc(1 - 2 + 3").get();
+
+  t(calculation1.isNumber());
+
+  t.deepEqual(calculation1.toJSON(), {
+    type: "math expression",
+    expression: {
+      type: "value",
+      value: { type: "number", value: 2 },
+    },
+  });
+
+  const calculation2 = parse("calc(10 / 2 * 5").get();
+
+  t(calculation2.isNumber());
+
+  t.deepEqual(calculation2.toJSON(), {
+    type: "math expression",
+    expression: {
+      type: "value",
+      value: { type: "number", value: 25 },
+    },
+  });
+});
+
+test(".parse() rejects sums without surrounding spaces", (t) => {
+  const calculation1 = parse("calc(1+2)");
+
+  t(calculation1.isErr());
+
+  const calculation2 = parse("calc(1-2)");
+
+  t(calculation2.isErr());
+});
+
+test(".parse() accepts products without surrounding spaces", (t) => {
+  const calculation1 = parse("calc(1*2)").get();
+
+  t(calculation1.isNumber());
+
+  t.deepEqual(calculation1.toJSON(), {
+    type: "math expression",
+    expression: {
+      type: "value",
+      value: { type: "number", value: 2 },
+    },
+  });
+
+  const calculation2 = parse("calc(1/2)").get();
+
+  t(calculation2.isNumber());
+
+  t.deepEqual(calculation2.toJSON(), {
+    type: "math expression",
+    expression: {
+      type: "value",
+      value: { type: "number", value: 0.5 },
+    },
+  });
+});
+
+test(".pares() accepts nesting parenthesis", (t) => {
+  const calculation = parse("calc((1 + 1) * (2 + 2))").get();
+
+  t(calculation.isNumber());
+
+  t.deepEqual(calculation.toJSON(), {
+    type: "math expression",
+    expression: {
+      type: "value",
+      value: { type: "number", value: 8 },
     },
   });
 });