+++ packages/alfa-style/test/property/font-size.spec.tsx
@@ -13,25 +13,22 @@ test("#cascaded() parses `font-size: calc(1em + 2px)`", (t) => {
 
   t.deepEqual(cascaded(element, "font-size"), {
     value: {
-      type: "calculation",
+      type: "math expression",
       expression: {
-        type: "sum",
-        operands: [
+        type: "calculation",
+        arguments: [
           {
-            type: "value",
-            value: {
-              type: "length",
-              value: 1,
-              unit: "em",
-            },
-          },
-          {
-            type: "value",
-            value: {
-              type: "length",
-              value: 2,
-              unit: "px",
-            },
+            type: "sum",
+            operands: [
+              {
+                type: "value",
+                value: { type: "length", value: 1, unit: "em" },
+              },
+              {
+                type: "value",
+                value: { type: "length", value: 2, unit: "px" },
+              },
+            ],
           },
         ],
       },
@@ -46,11 +43,7 @@ test("#computed() resolves `font-size: calc(1em + 2px)`", (t) => {
   const style = Style.from(element, device);
 
   t.deepEqual(style.computed("font-size").toJSON(), {
-    value: {
-      type: "length",
-      value: 18,
-      unit: "px",
-    },
+    value: { type: "length", value: 18, unit: "px" },
     source: h.declaration("font-size", "calc(1em + 2px)").toJSON(),
   });
 });
@@ -60,24 +53,22 @@ test("#cascaded() parses `font-size: calc(1em + 2%)`", (t) => {
 
   t.deepEqual(cascaded(element, "font-size"), {
     value: {
-      type: "calculation",
+      type: "math expression",
       expression: {
-        type: "sum",
-        operands: [
+        type: "calculation",
+        arguments: [
           {
-            type: "value",
-            value: {
-              type: "length",
-              value: 1,
-              unit: "em",
-            },
-          },
-          {
-            type: "value",
-            value: {
-              type: "percentage",
-              value: 0.02,
-            },
+            type: "sum",
+            operands: [
+              {
+                type: "value",
+                value: { type: "length", value: 1, unit: "em" },
+              },
+              {
+                type: "value",
+                value: { type: "percentage", value: 0.02 },
+              },
+            ],
           },
         ],
       },
@@ -92,11 +83,7 @@ test("#computed() resolves `font-size: calc(1em + 2%)`", (t) => {
   const style = Style.from(element, device);
 
   t.deepEqual(style.computed("font-size").toJSON(), {
-    value: {
-      type: "length",
-      value: 16.32,
-      unit: "px",
-    },
+    value: { type: "length", value: 16.32, unit: "px" },
     source: h.declaration("font-size", "calc(1em + 2%)").toJSON(),
   });
 });
@@ -115,11 +102,7 @@ test("#computed() resolves `font-size: calc(100% + 2%)`", (t) => {
   const style = Style.from(element, device);
 
   t.deepEqual(style.computed("font-size").toJSON(), {
-    value: {
-      type: "length",
-      value: 16.32,
-      unit: "px",
-    },
+    value: { type: "length", value: 16.32, unit: "px" },
     source: h.declaration("font-size", "calc(100% + 2%)").toJSON(),
   });
 });